Move ProjectionErrorMeasurement to a separate file
